refac: 테스트 코드와 주요 로직 패키지 분리
