+++ src/test/kotlin/study/DslTest.kt
@@ -1,5 +1,6 @@
 package study
 
+import dsl.introduce
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
@@ -30,20 +31,5 @@ class DslTest {
 
         person.name shouldBe name
         person.company shouldBe company
-    }
-}
-
-fun introduce(block: Person.() -> Unit): Person = Person().apply(block)
-
-class Person() {
-    lateinit var name: String
-    lateinit var company: String
-
-    fun name(value: String) {
-        name = value
-    }
-
-    fun company(value: String) {
-        company = value
     }
 }
